+++ scripts/generate-app-approval.mjs
@@ -1,6 +1,6 @@
 'use strict'
 
 import fs from 'fs'
-import { _Connectors } from '../blockchain/build/index.main.mjs'
+import { _Connectors } from '../blockchain/token-market/build/index.main.mjs'
 
-fs.writeFileSync('./blockchain/build/app-approval.mjs', `export const appApproval = '${_Connectors.ALGO.appApproval}'`)
+fs.writeFileSync('./blockchain/token-market/build/app-approval.mjs', `export const appApproval = '${_Connectors.ALGO.appApproval}'`)
